fix(builders): handle isolated and Windows builds
